Extract parquet logic into submodule from datafusion_core::dataframe
