+++ datafusion/core/src/datasource/mod.rs
@@ -43,4 +43,5 @@ pub use self::provider::TableProvider;
 pub use self::view::ViewTable;
 pub use crate::logical_expr::TableType;
 #[cfg(feature = "parquet")]
-pub(crate) use statistics::{get_col_stats, get_statistics_with_limit};
+pub(crate) use statistics::get_col_stats;
+pub use statistics::get_statistics_with_limit;
